specìfication payment
CRUD specification

+++ Backend/src/main/java/com/egustore/eshop/repository/PaymentRepository.java
@@ -1,7 +1,8 @@
 package com.egustore.eshop.repository;
 
-import com.egustore.eshop.entity.Payment;
+import com.egustore.eshop.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 }
